arm/arm64: Change dcache_line_size to ulong
dcache_line_size is treated like a long in assembly, so make it one.

Message-Id: <20201014191444.136782-4-drjones@redhat.com>

+++ lib/arm/asm/processor.h
@@ -89,6 +89,6 @@ static inline u32 get_ctr(void)
 	return read_sysreg(CTR);
 }
 
-extern u32 dcache_line_size;
+extern unsigned long dcache_line_size;
 
 #endif /* _ASMARM_PROCESSOR_H_ */
